added demo folder, added Markdown with call to Synapse
